Rename Main.js to Authentication.js
