+++ app/javascript/components/homepage/Home.js
@@ -3,8 +3,8 @@ import { BrowserRouter as Router } from 'react-router-dom';
 import { Route } from 'react-router-dom';
 import NaviBar from './NaviBar';
 import HomeContainer from './HomeContainer';
-import LoginForm from '../users/main/LoginForm';
-import SignUpForm from '../users/main/SignUpForm';
+import LoginForm from '../users/authentication/LoginForm';
+import SignUpForm from '../users/authentication/SignUpForm';
 
 export default function Home() {
     return (
