+++ app/javascript/components/users/authentication/SignUpForm.js
