--- app/views/users/main.html.erb
@@ -1,1 +1,0 @@
-<%= react_component("users/main/Main") %>
