+++ app/views/users/sessions/new.html.erb
@@ -1,1 +1,1 @@
-<%= react_component("users/main/Main") %>
+<%= react_component("users/Authentication") %>
